new_codes/requirements.txt
# Core dependencies for RAG Pipeline
streamlit>=1.28.0
openai>=1.0.0
pymilvus>=2.3.0
python-dotenv>=1.0.0

# Document processing
pdfplumber>=0.10.0
python-docx>=1.0.0

# Text processing
tiktoken>=0.5.0

# ML/AI dependencies
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0

